ext/pkg/io/read_test.go
package io

import (
	"bytes"
	"context"
	"io"
	"os"
	"testing"
	"time"

	"github.com/go-faker/faker/v4"
	"github.com/siyul-park/uniflow/pkg/node"
	"github.com/siyul-park/uniflow/pkg/object"
	"github.com/siyul-park/uniflow/pkg/packet"
	"github.com/siyul-park/uniflow/pkg/port"
	"github.com/siyul-park/uniflow/pkg/process"
	"github.com/stretchr/testify/assert"
)

func TestNewReadNode(t *testing.T) {
	f, _ := os.CreateTemp("", "*")
	defer f.Close()

	n := NewReadNode(f)
	assert.NotNil(t, n)
	assert.NoError(t, n.Close())
}

func TestReadNode_SendAndReceive(t *testing.T) {
	ctx, cancel := context.WithTimeout(context.TODO(), time.Second)
	defer cancel()

	contents := []byte(faker.Sentence())
	r := bytes.NewReader(contents)

	n := NewReadNode(io.NopCloser(r))
	defer n.Close()

	in := port.NewOut()
	in.Link(n.In(node.PortIn))

	proc := process.New()
	defer proc.Exit(nil)

	inWriter := in.Open(proc)

	inPayload := object.NewInt(len(contents))
	inPck := packet.New(inPayload)

	inWriter.Write(inPck)

	select {
	case outPck := <-inWriter.Receive():
		assert.Equal(t, object.NewBinary(contents), outPck.Payload())
	case <-ctx.Done():
		assert.Fail(t, ctx.Err().Error())
	}
}

func TestReadNodeCodec_Decode(t *testing.T) {
	codec := NewReadNodeCodec()

	spec := &ReadNodeSpec{
		Filename: "stdin",
	}

	n, err := codec.Decode(spec)
	assert.NoError(t, err)
	assert.NotNil(t, n)
	assert.NoError(t, n.Close())
}
